requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
motor==3.7.1
pydantic[email]==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
celery==5.6.0
redis==6.4.0
fastapi-mail==1.5.8
loguru==0.7.2
pillow==12.1.0
httpx==0.28.1
itsdangerous==2.2.0
slowapi==0.1.9